general_llm/requirements.txt
llama-index
llama-parse
pinecone-client[grpc]
arxiv
setuptools
llama-index-llms-ollama
python-dotenv
llama-index-vector-stores-pinecone
llama-index-embeddings-huggingface